Extracted FormulaHelper to separate file
