+++ test/table_test.rb
@@ -435,14 +435,4 @@ class TableTest < Minitest::Test
     assert_equal("Column 'Name' already exists", e.to_s)
 
   end
-
-  def test_create_fomula
-    f = TableTransform::Formula
-    assert_equal('"no"', f::text('no'))
-    assert_equal('[C1]', f::column('C1'))
-    assert_equal('T1[]', f::table('T1'))
-    assert_equal('VLOOKUP(C1,T1[],COLUMN(T1[[#Headers],[R1]]),FALSE)', f::vlookup('C1', 'T1', 'R1'))
-    assert_equal('VLOOKUP([C1],T1[],COLUMN(T1[[#Headers],[R1]]),FALSE)', f::vlookup(f::column('C1'), 'T1', 'R1'))
-
-  end
 end
